scripts/support/requirements.txt
gitpython >= 2.5.1
